+++ cmd/common.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"encoding/json"
-	"fmt"
 	"log"
 	"os"
 	"strings"
@@ -63,16 +62,6 @@ func newConditionsClient(ctx context.Context, mctl *app.App) (*co.Client, error)
 	)
 }
 
-func findAttribute(ns string, attributes []serverservice.Attributes) *serverservice.Attributes {
-	for _, attribute := range attributes {
-		if attribute.Namespace == ns {
-			return &attribute
-		}
-	}
-
-	return nil
-}
-
 func AttributeFromLabels(ns string, labels map[string]string) (*serverservice.Attributes, error) {
 	data, err := json.Marshal(labels)
 	if err != nil {
@@ -83,13 +72,4 @@ func AttributeFromLabels(ns string, labels map[string]string) (*serverservice.At
 		Namespace: ns,
 		Data:      data,
 	}, nil
-}
-
-func printJSON(data interface{}) {
-	b, err := json.MarshalIndent(data, "", "  ")
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println(string(b))
 }
