+++ cmd/firmware.go
@@ -1,62 +1,16 @@
 package cmd
 
 import (
-	"context"
 	"encoding/json"
 	"log"
-	"os"
-	"strings"
-	"time"
 
 	"github.com/google/uuid"
 	"github.com/metal-toolbox/mctl/internal/app"
-	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
 
 	coApi "github.com/metal-toolbox/conditionorc/pkg/api/v1/types"
 	coTyp "github.com/metal-toolbox/conditionorc/pkg/types"
 )
-
-var (
-	cmdTimeout = 20 * time.Second
-)
-
-// List
-var cmdListFirmware = &cobra.Command{
-	Use:   "firmware",
-	Short: "List firmware",
-	Run: func(cmd *cobra.Command, args []string) {
-		mctl, err := app.New(cmd.Context(), cfgFile)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		ctx, cancel := context.WithTimeout(cmd.Context(), cmdTimeout)
-		defer cancel()
-
-		c, err := newServerserviceClient(ctx, mctl)
-		if err != nil {
-			log.Fatal("error initializing serverservice client: ", err)
-		}
-
-		firmware, _, err := c.ListServerComponentFirmware(cmd.Context(), nil)
-		if err != nil {
-			log.Fatal("serverservice client returned error: ", err)
-		}
-
-		if outputJSON {
-			printJSON(firmware)
-			os.Exit(0)
-		}
-
-		table := tablewriter.NewWriter(os.Stdout)
-		table.SetHeader([]string{"UUID", "Vendor", "Model", "Component", "Version"})
-		for _, f := range firmware {
-			table.Append([]string{f.UUID.String(), f.Vendor, strings.Join(f.Model, ","), f.Component, f.Version})
-		}
-		table.Render()
-	},
-}
 
 type installParams struct {
 	FirmwareID uuid.UUID `json:"firmwareId"`
