+++ main.go
@@ -21,6 +21,7 @@ import (
 	_ "github.com/metal-toolbox/mctl/cmd/delete"
 	_ "github.com/metal-toolbox/mctl/cmd/edit"
 	_ "github.com/metal-toolbox/mctl/cmd/get"
+	_ "github.com/metal-toolbox/mctl/cmd/list"
 )
 
 func main() {
